3.OCR/Al_resume_project/requirements.txt
streamlit
pdfplumber
spacy
genai
pandas
scikit-learn
sentence-transformers
tiktoken>=0.5.1
google-generativeai
python-dotenv
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
